Certificate-Sender/requirements.txt
pandas==1.2.4
openpyxl==3.0.7
Pillow==8.2.0